(multiplayer) Changed my mind, so sue me!  Still not happy with this, but it is only demo code, not going to go crazy on it.

+++ C++/MultiPlayer/Step 4 - Entity Prediction/shared/components/Movement.hpp
@@ -15,14 +15,14 @@ namespace components
     class Movement : public Component
     {
       public:
-        Movement(double thrustRate, float rotateRate, math::Vector2f momentum) :
+        Movement(float thrustRate, float rotateRate, math::Vector2f momentum) :
             m_thrustRate(thrustRate),
             m_rotateRate(rotateRate),
             m_momentum(momentum)
         {
         }
 
-        const double getThrustRate() const { return m_thrustRate; }
+        const float getThrustRate() const { return m_thrustRate; }
         const float getRotateRate() const { return m_rotateRate; }
         const math::Vector2f getMomentum() const { return m_momentum; }
         void setMomentum(math::Vector2f momentum) { m_momentum = momentum; }
@@ -32,9 +32,9 @@ namespace components
         void updateIntraMovementTime(std::chrono::microseconds howMuch) { m_intraMovementTime += howMuch; }
 
       private:
-        double m_thrustRate;       // unit distance per microsecond
-        float m_rotateRate;        // degrees per microsecond
-        math::Vector2f m_momentum; // units per microsecond
+        float m_thrustRate;       // unit distance per millisecond
+        float m_rotateRate;        // degrees per millisecond
+        math::Vector2f m_momentum; // units per millisecond
         std::chrono::microseconds m_intraMovementTime{0};
     };
 } // namespace components
