+++ C++/MultiPlayer/Step 5 - Weapons/shared/entities/Player.hpp
@@ -15,7 +15,7 @@
 // --------------------------------------------------------------
 namespace entities::player
 {
-    std::shared_ptr<Entity> create(std::string texture, math::Vector2f position, float size, double thrustRate, float rotateRate, math::Vector2f momentum);
+    std::shared_ptr<Entity> create(std::string texture, math::Vector2f position, float size, float thrustRate, float rotateRate, math::Vector2f momentum);
 }
 
 namespace entities
